+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## in development
 
+### 2019-01-04
+* Objects, Blocks, Threads: new feature/block: sense colors and sprites anywhere 
+
 ### 2019-01-02
 * Objects, Blocks, Threads, GUI, Store: added third color dimension and transparency to pen
 * renamed help screens for "setHue" to "setPenHSVA" and "changeHue" to "changePenHSVA"